wip(test_signals): make some fixtures available to others
